Move networking off main thread... and stuff.
These changes are required to support deploying NanoHTTPD on the Android
platform.  Here we move the creation of the server socket into the
serving thread, rather than doing it outside.

This change also adds a monitor to synchronize startup, so that upon
exiting this method we know the thread has definitely begun execution.

Next, I've added a couple checks to make it at least a little safer to
call serve() multiple times (although we need extra locking to make this
truly re-entrant).

Lastly, I've moved the server socket creation code into a separate
protected method.  This makes it possible for users of the class to
override the stock socket creation logic.