Reorganize serve methods.
The method serve(IHTTPSession) was not useful to override on its own as
the stock version executed a bunch of very important logic in it (parsing out
request bodies and so forth).  You could call super.serve() and throw
away the Response object, but that feels like a serious hack.

So here I've taken that core logic and moved it into a method called
serveInternal.  That method does its thing and then passes control to
serve().