Add hook for parsing "unusual" POST bodies.
It's possible to see a POST with a content-type that isn't one of the
standard form types (I'm thinking of JSON, here).  This change makes it
possible for users of the class to hook in and provide their own body
parsing logic (like deserializing a JSON string into an object).